debug

+++ shtdlib.sh
@@ -2310,12 +2310,15 @@ function test_add_on_mod {
     add_on_mod "signal_process ${signaler_pid} SIGUSR1 &> /dev/null" "${tmp_file_path}" &
     mod_watcher_pid="${!}"
     bash -c "sleep 2 && echo 'test message' > '${tmp_file_path}'"
-    ps -Afl
     bash -c "sleep 10 && kill ${signaler_pid} &> /dev/null" &
     while pgrep -P ${$} > /dev/null ; do
         wait ${signaler_pid} &> /dev/null
         # Make sure the sub process exits with 42
-        assert [ "${?}" == '42' ]
+        return_status="${?}"
+        if [ "${return_status}" != '42' ] ; then
+            debug 1 "Got return status ${return_status} when waiting for PID ${signaler_pid} to exit"
+            exit_on_fail
+        fi
         color_echo green "Sub process was signaled by file system monitoring thread, responded and properly exited"
         debug 10 "Signaling mod_watcher ${mod_watcher_pid} to exit"
         kill "${mod_watcher_pid}"
